+++ prime-factors/PrimeFactorsApp-02/PrimeFactors/Sieve.cs
@@ -1,17 +1,17 @@
 ﻿namespace PrimeFactors;
 
 public class Sieve : PrimeGenerator {
-  public IEnumerable<int> GetPrimesUpTo(int n)
-    => n < 2 ? Enumerable.Empty<int>() : PrimesUpTo(n);
+  public IEnumerable<long> GetPrimesUpTo(long n)
+    => n < 2 ? Enumerable.Empty<long>() : PrimesUpTo(n);
 
-  static IEnumerable<int> PrimesUpTo(int n) {
-    List<int> primes = new();
+  static IEnumerable<long> PrimesUpTo(long n) {
+    List<long> primes = new();
     bool[] composites = new bool[n + 1];
 
-    for (int candidate = 2; candidate <= n; candidate++) {
+    for (long candidate = 2; candidate <= n; candidate++) {
       if (composites[candidate] == false) {
         primes.Add(candidate);
-        for (int multiple = 2 * candidate; multiple <= n; multiple += candidate) {
+        for (long multiple = 2 * candidate; multiple <= n; multiple += candidate) {
           composites[multiple] = true;
         }
       }
